Fix Integer Overflow (#7)
* Fix Integer Overflow

When reading from a very large file the references to bookkeeping
how many bytes are remaining and how many bytes have been read from
the file would overflow causing the buffer limit to be set to zero.

When the buffer limit is set to zero the channel wouldn't read any
more bytes from the file and the reader would effectively spin-lock
at the current position and halt any progress on reading the file.

The result of the spinlock would be a runaway process that would
never get killed. A client requesting the lines would hang up
on the request since the server never returned any data.

This commit updates the bookkeeping references in the reader
implementation to be a long type instead of int. Furthermore,
it refactors the logic to limit the buffer when at the head of the
file to safely cast the new limit to an integer.

* Add logging

- An access logger following standard apache log format
  - Extended to format to show the response time in milliseconds.
- Implementation logging so that we can see file reading durations.